feat: automatically insert seed word from clipboard when importing
If the clipboard text contains a valid seed word list, the inputs are
automatically filled with its content.

To be a valid seed word list we perform 3 checks:

- word list from clipboard match seed size (12, 15, 18, 21, or 24 words)
- all words are valid from the BIP39 word list
- checksum passes

The text content of the clipboard is split by both spaces and newlines
and the resulting array is checked upon the above criteria.