fix: assume type 'send' for transactions which receive no funds
Given a raw transaction, we can know how much an address has received
because we know the outputs' values and the outputs' addresses, but we
cannot know how much has been spent by the address in that transaction
because we need to check if the transaction inputs belong to that
address. Therefore, it is a reasonable heuristic to assume the
transaction is of type 'send' if the received amount is 0.