chore: add rules to gitignore
By default, the Android SDK, Expo and Yarn create hidden files in the
home directory, but this can turn into a mess if the developer is
working under multiple projects. Changing the HOME env variable to the
current working directory when working inside a project will make those
programs create the hidden files in the current project, which does not
mess up with other projects.