fix: inserting PIN from virtual keyboard not working
It did not work due to race condition: the `handleLastPin` was always
called before the state variable `pin` had been updated. It happened
because the event `onChangeText` was triggered after the event
`onKeyPress`. The `onChangeText` was responsible for updating the `pin`
state variable, while `onKeyPress` was responsible for calling
`handleLastPin`. The race condition has been solved by refactoring code
such that the `handleLastPin` is called by `onChangeText` and the
updated `pin` variable is passed as an argument to that function.